Add files via upload
This project is about Support vector Machine and its Application.In this Application we did Digit recognition using Sklearn Library in Python.SVM classifier is used as the training algorithm and then tested it on MNIST dataset. We achieved a training accuracy of 98.05% and a test accuracy of 97.83% demonstrating that the proposed method can achieve significant and promising performance in digit recognition. Then we implemented our model to recognize user given handwritten digits in real time.SVM has Some kernel function .In this digit recognition we use radic basis Function . Radial basis function is typically the best option (RBF). Currently, the proper kernel function is chosen using the cross validation method.